extracted wrongly placed functionality from Graph and then put it inside the Traversals class.
