+++ src/edu/ucsc/refactor/util/graph/Graph.java
@@ -8,20 +8,12 @@ import java.util.*;
  * @author hsanchez@cs.ucsc.edu (Huascar A. Sanchez)
  */
 public class Graph<T> {
-    /** Color used to mark unvisited nodes */
-    public static final int VISIT_COLOR_WHITE = 1;
-
-    /** Color used to mark nodes as they are first visited in DFS order */
-    public static final int VISIT_COLOR_GREY = 2;
-
-    /** Color used to mark nodes after descendants are completely visited */
-    public static final int VISIT_COLOR_BLACK = 3;
 
     /** Vector<Vertex> of graph vertices */
-    private List<Vertex<T>> vertices;
+    private final List<Vertex<T>> vertices;
 
     /** Vector<Edge> of edges in the graph */
-    private List<Edge<T>> edges;
+    private final List<Edge<T>> edges;
 
     /** The vertex identified as the root of the graph */
     private Vertex<T> rootVertex;
@@ -31,7 +23,7 @@ public class Graph<T> {
      */
     public Graph() {
         vertices = new ArrayList<Vertex<T>>();
-        edges = new ArrayList<Edge<T>>();
+        edges    = new ArrayList<Edge<T>>();
     }
 
     /**
@@ -83,56 +75,16 @@ public class Graph<T> {
 
 
     /**
-     * Perform a breadth first search of this graph, starting at v.
+     * Set a root vertex. If root does no exist in the graph it is added.
      *
-     * @param v -
-     *          the search starting point
-     * @param visitor -
-     *          the visitor whose visit method is called prior to visiting a vertex.
+     * @param root -
+     *          the vertex to set as the root and optionally add if it does not
+     *          exist in the graph.
      */
-    public void breadthFirstSearch(Vertex<T> v, final Visitor<T> visitor) {
-        VisitorThrow<T, RuntimeException> wrapper = new VisitorThrow<T, RuntimeException>() {
-            public void visit(Graph<T> g, Vertex<T> v) throws RuntimeException {
-                if (visitor != null)
-                    visitor.visit(g, v);
-            }
-        };
-        this.breadthFirstSearch(v, wrapper);
-    }
-
-    /**
-     * Perform a breadth first search of this graph, starting at v. The visit may
-     * be cut short if visitor throws an exception during a visit callback.
-     *
-     * @param <E>
-     * @param v -
-     *          the search starting point
-     * @param visitor -
-     *          the visitor whose visit method is called prior to visiting a vertex.
-     * @throws E
-     *           if visitor.visit throws an exception
-     */
-    public <E extends Exception> void breadthFirstSearch(Vertex<T> v, VisitorThrow<T, E> visitor)
-            throws E {
-        LinkedList<Vertex<T>> q = new LinkedList<Vertex<T>>();
-
-        q.add(v);
-        if (visitor != null)
-            visitor.visit(this, v);
-        v.visit();
-        while (!q.isEmpty()) {
-            v = q.removeFirst();
-            for (int i = 0; i < v.getOutgoingEdgeCount(); i++) {
-                Edge<T> e = v.getOutgoingEdge(i);
-                Vertex<T> to = e.getTo();
-                if (!to.visited()) {
-                    q.add(to);
-                    if (visitor != null)
-                        visitor.visit(this, to);
-                    to.visit();
-                }
-            }
-        }
+    public void addRootVertex(Vertex<T> root) {
+        this.rootVertex = root;
+        if (!containsVertex(root))
+            this.addVertex(root);
     }
 
 
@@ -176,51 +128,6 @@ public class Graph<T> {
 
 
     /**
-     * Perform a depth first search using recursion.
-     *
-     * @param v -
-     *          the Vertex to start the search from
-     * @param visitor -
-     *          the visitor to inform prior to
-     * @see Visitor#visit(Graph, Vertex)
-     */
-    public void depthFirstSearch(Vertex<T> v, final Visitor<T> visitor) {
-        VisitorThrow<T, RuntimeException> wrapper = new VisitorThrow<T, RuntimeException>() {
-            public void visit(Graph<T> g, Vertex<T> v) throws RuntimeException {
-                if (visitor != null)
-                    visitor.visit(g, v);
-            }
-        };
-        this.depthFirstSearch(v, wrapper);
-    }
-
-    /**
-     * Perform a depth first search using recursion. The search may be cut short
-     * if the visitor throws an exception.
-     *
-     * @param <E>
-     * @param v -
-     *          the Vertex to start the search from
-     * @param visitor -
-     *          the visitor to inform prior to
-     * @see Visitor#visit(Graph, Vertex)
-     * @throws E
-     *           if visitor.visit throws an exception
-     */
-    public <E extends Exception> void depthFirstSearch(Vertex<T> v, VisitorThrow<T, E> visitor) throws E {
-        if (visitor != null)
-            visitor.visit(this, v);
-        v.visit();
-        for (int i = 0; i < v.getOutgoingEdgeCount(); i++) {
-            Edge<T> e = v.getOutgoingEdge(i);
-            if (!e.getTo().visited()) {
-                depthFirstSearch(e.getTo(), visitor);
-            }
-        }
-    }
-
-
-    /**
      * Search the vertices for one with name.
      *
      * @param name -
@@ -258,32 +165,6 @@ public class Graph<T> {
         return match;
     }
 
-    /**
-     * Search the graph for cycles. In order to detect cycles, we use a modified
-     * depth first search called a colored DFS. All nodes are initially marked
-     * white. When a node is encountered, it is marked grey, and when its
-     * descendants are completely visited, it is marked black. If a grey node is
-     * ever encountered, then there is a cycle.
-     *
-     * @return the edges that form cycles in the graph. The array will be empty if
-     *         there are no cycles.
-     */
-    public Edge<T>[] findCycles() {
-        ArrayList<Edge<T>> cycleEdges = new ArrayList<Edge<T>>();
-        // Mark all vertices as white
-        for (int n = 0; n < vertices.size(); n++) {
-            Vertex<T> v = getVertex(n);
-            v.setMarkState(VISIT_COLOR_WHITE);
-        }
-        for (int n = 0; n < vertices.size(); n++) {
-            Vertex<T> v = getVertex(n);
-            visit(v, cycleEdges);
-        }
-
-        @SuppressWarnings("unchecked") Edge<T>[] cycles = new Edge[cycleEdges.size()];
-        cycleEdges.toArray(cycles);
-        return cycles;
-    }
 
     /**
      * Get the root vertex
@@ -452,36 +333,6 @@ public class Graph<T> {
         return vertices.size();
     }
 
-    /**
-     * Set a root vertex. If root does no exist in the graph it is added.
-     *
-     * @param root -
-     *          the vertex to set as the root and optionally add if it does not
-     *          exist in the graph.
-     */
-    public void setRootVertex(Vertex<T> root) {
-        this.rootVertex = root;
-        if (!containsVertex(root))
-            this.addVertex(root);
-    }
-
-
-    private void visit(Vertex<T> v, ArrayList<Edge<T>> cycleEdges) {
-        v.setMarkState(VISIT_COLOR_GREY);
-        int count = v.getOutgoingEdgeCount();
-        for (int n = 0; n < count; n++) {
-            Edge<T> e = v.getOutgoingEdge(n);
-            Vertex<T> u = e.getTo();
-            if (u.getMarkState() == VISIT_COLOR_GREY) {
-                // A cycle Edge<T>
-                cycleEdges.add(e);
-            } else if (u.getMarkState() == VISIT_COLOR_WHITE) {
-                visit(u, cycleEdges);
-            }
-        }
-
-        v.setMarkState(VISIT_COLOR_BLACK);
-    }
 
     @Override public String toString() {
         StringBuilder tmp = new StringBuilder("Graph[");
